main.py
It is another utility app in which the user can set a timer, and the app notifies you when the time is up.

This nifty app helps in furthering knowledge of python coding. It is an intermediary-level project in which the codes will take gather the input corresponding to the length of the countdown within seconds. After receiving the input, the countdown will be initiated and will appear on the screen in “minutes: seconds” format.